Remove 'configuration' from class names because the search is no restricted to configuration files only.
ConfigurationSearcher > FileSearcher
ConfigFileHit > FileHit
