+++ source/GitHubSearch.csproj
@@ -62,9 +62,9 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Common\Bootstrapper.cs" />
-    <Compile Include="ConfigFileResult.cs" />
+    <Compile Include="FileHit.cs" />
     <Compile Include="Configuration.cs" />
-    <Compile Include="ConfigurationSearcher.cs" />
+    <Compile Include="FileSearcher.cs" />
     <Compile Include="FileCache.cs" />
     <Compile Include="GitHubAdapter.cs" />
     <Compile Include="GitHubClientFactory.cs" />
